Integrate ellipse functionality into geo_core
- Consolidate ellipse calculations from analysis/curves.rs and geo_core/metrics/ellipse.rs into geo_core/approximations/ellipse.rs
- Implement all ellipse methods with Scalar generics for type safety
- Add comprehensive test suite for ellipse functions
- Remove duplicate files: geo_core/metrics/ellipse.rs
- Convert analysis/curves.rs to migration notice with documentation
- Maintain Foundation Pattern compliance
- Add convenient re-exports in geo_core/lib.rs

Features unified:
- Ramanujan approximations (I & II)
- Series expansion method
- Numerical integration method
- Padé and Cantrell approximations
- Ellipse properties (area, eccentricity, focal distance, foci)
- All with Scalar<T> generic support

Resolves ellipse function duplication across analysis and geo_core modules
